feat: Add consolidated version of fragment
