--- src/components/Log/components/index.tsx
@@ -1,3 +1,0 @@
-export { default as LogMovie } from "./LogMovie";
-export { default as LogTV } from "./LogTV";
-export { default as LogSpotify } from "./LogSpotify";
