+++ src/components/Media/components/MediaEditEpisodes.tsx
@@ -1,0 +1,65 @@
+import {
+  Text,
+  Flex,
+  Button,
+  Divider,
+  SimpleGrid,
+  Checkbox,
+} from "@chakra-ui/react";
+import React, { useState } from "react";
+import type { Dispatch } from "react";
+import type { LogActions, LogState } from "../../../config/storeLog";
+
+interface Props {
+  dispatch: Dispatch<LogActions>;
+  fields: LogState;
+  episodes: number;
+}
+export default function MediaEditEpisodes({
+  fields,
+  dispatch,
+  episodes,
+}: Props): JSX.Element {
+  const { seenEpisodes } = fields;
+  return (
+    <>
+      <Divider my={2} />
+      <SimpleGrid columns={Math.floor(episodes / 3)} spacingY={3}>
+        {Array.from({ length: episodes }, (_, episodeNumber: number) => (
+          <Checkbox
+            key={`episode_${episodeNumber + 1}`}
+            value={episodeNumber + 1}
+            isChecked={seenEpisodes?.includes(episodeNumber + 1)}
+            colorScheme="purple"
+            onChange={(e) => {
+              // TODO: prevent empty episode number
+              let episodeArr: number[] = [];
+              if (typeof seenEpisodes !== "undefined") {
+                const hasValue = seenEpisodes.includes(
+                  parseInt(e.target.value)
+                );
+                episodeArr = hasValue
+                  ? seenEpisodes.filter(
+                      (filterValue) => filterValue !== parseInt(e.target.value)
+                    )
+                  : [parseInt(e.target.value)].concat(seenEpisodes);
+              } else {
+                episodeArr = [parseInt(e.target.value)];
+              }
+
+              return dispatch({
+                type: "state",
+                payload: {
+                  key: "seenEpisodes",
+                  value: episodeArr,
+                },
+              });
+            }}
+          >
+            <Text fontSize="sm">{episodeNumber + 1}</Text>
+          </Checkbox>
+        ))}
+      </SimpleGrid>
+    </>
+  );
+}
