+++ src/components/Media/components/index.tsx
@@ -1,0 +1,12 @@
+export { default as MediaHeader } from "./MediaHeader";
+export { default as MediaContainer } from "./MediaContainer";
+export { default as MediaPoster } from "./MediaPoster";
+export { default as MediaInfo } from "./MediaInfo";
+export { default as MediaInfoText } from "./MediaInfoText";
+export { default as MediaInfoButton } from "./MediaInfoButton";
+export { default as MediaAbout } from "./MediaAbout";
+export { default as MediaCast } from "./MediaCast";
+export { default as MediaRating } from "./MediaRating";
+export { default as MediaEdit } from "./MediaEdit";
+export { default as MediaEditEpisodes } from "./MediaEditEpisodes";
+export { default as MediaEditSeasons } from "./MediaEditSeasons";
