+++ src/components/Media/index.tsx
@@ -1,0 +1,3 @@
+export { default as MediaMovie } from "./MediaMovie";
+export { default as MediaTV } from "./MediaTV";
+export { default as MediaSpotify } from "./MediaSpotify";
