+++ src/components/Selected/Selected.tsx
@@ -15,7 +15,7 @@ import type { SpotifyAlbum, SpotifyArtist } from "../../types/typesSpotify";
 import { parsePosterUrl } from "../../utils/helpers";
 import BookmarkIcon from "../icons/BookmarkIcon";
 import MdLoader from "../md/MdLoader";
-import { SelectedSpotify, SelectedMovie, SelectedTV } from "./components";
+import { MediaMovie, MediaSpotify, MediaTV } from "../Media";
 
 export default function Selected(): JSX.Element {
   const { user } = useFuegoUser();
@@ -88,13 +88,13 @@ function SelectedSuspense({
     <>
       <DrawerBody px={{ base: 6, sm: 8 }}>
         {item.type === "album" && (
-          <SelectedSpotify
+          <MediaSpotify
             artistInfo={(data as DataFetchSpotify)[1]}
             albumInfo={(data as DataFetchSpotify)[0]}
           />
         )}
-        {item.type === "movie" && <SelectedMovie data={data as MDbMovie} />}
-        {item.type === "tv" && <SelectedTV data={data as MDbTV} />}
+        {item.type === "movie" && <MediaMovie data={data as MDbMovie} />}
+        {item.type === "tv" && <MediaTV data={data as MDbTV} />}
       </DrawerBody>
       <DrawerFooter
         borderTopWidth="1px"
