--- src/components/Selected/components/index.tsx
@@ -1,3 +1,0 @@
-export { default as SelectedMovie } from "./SelectedMovie";
-export { default as SelectedTV } from "./SelectedTV";
-export { default as SelectedSpotify } from "./SelectedSpotify";
